requirement.txt
click>=7.1.2
patch
pytest
attrs==19.1.0
pytest-celery
